[Google_places_api] Create new classes, made refactoring
